requirements.txt
asgiref==3.9.0
certifi==2025.6.15
channels==4.2.2
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
pillow==11.3.0
regex==2024.11.6
requests==2.32.4
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
